requirements.txt
gunicorn
# Backend requirements (mirrored from backend/requirements.txt)
# Core Django Framework
celery
Django==5.2.4
djangorestframework==3.16.0

# Database
psycopg2-binary==2.9.10

# API and CORS
django-cors-headers==4.7.0
djangorestframework-camel-case==1.4.2

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# AI and Chatbot Dependencies (Pinned to local environment)
google-generativeai==0.3.2
langchain
google-genai
langchain-community
langchain-groq
python-decouple
python-dotenv
cryptography

# Database
sqlalchemy==2.0.35

# Graph Database (if using Neo4j features)
neomodel==5.3.3
django-neomodel==0.2.0

# HTTP Requests (for testing and external API calls)
requests==2.31.0

# Payment Gateways
razorpay==1.3.0

# Google Play Billing (for TWA in-app purchases)
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Redis for OTP caching and rate limiting
redis==5.0.1

# SMS via MSG91 (replaced AWS SNS)
# MSG91 uses standard Python http.client (no additional packages needed)
# boto3==1.35.9  # <-- Removed: Previously used for AWS SNS
# botocore==1.35.9  # <-- Removed: Previously used for AWS SNS

# Excel file processing for institution uploads
openpyxl==3.1.2

# Image processing (for base64 image validation, compression, PDF generation)
Pillow==10.1.0

# Development and Testing Dependencies
# (Optional - can be moved to separate requirements-dev.txt)
# These are used in test files but not required for production

spacy==3.7.4
pydantic==2.7.4
razorpay
